Transfer changes to download.py
